Added documentation to modules and opted for more specific integer types from stdint.

+++ include/neighbourhoods.h
@@ -10,32 +10,39 @@
 #include "environment.h"
 #include <stdbool.h>
 
+/** Represents a coordinate in the 2D plane. */
 typedef struct coord {
-    int x;
-    int y;
+    int32_t x; /**< The x component of the coordinate. */
+    int32_t y; /**< The y component of the coordinate. */
 } Coordinate;
 
+/** Represents a cell's neighbourhood. */
 typedef struct neighbourhood {
-    unsigned short int size;
-    Coordinate neighbours[];
+    uint8_t size;            /**< The number of grid cells in the neighbourhood. */
+    Coordinate neighbours[]; /**< The locations of each neighbour in relative vectors from the cell being considered. */
 } Neighbourhood;
 
-/* NEIGHBOURHOODS */
+/** Von Neumann neighbourhood. */
 #define VonNeumann                                                                                                     \
     {0, -1}, {0, 1}, {1, 0}, { -1, 0 }
+/** Von Neumann neighbourhood (just the corners). */
 #define VonNeumannCorners                                                                                              \
     {1, -1}, {1, 1}, {-1, -1}, { -1, 1 }
+/** Lesse neighbourhood. */
 #define Lesse                                                                                                          \
     VonNeumannCorners, {0, -2}, {0, 2}, {2, 0}, { -2, 0 }
+/** Moore neighbourhood. */
 #define Moore VonNeumann, VonNeumannCorners
+/** Von Neumann R2 neighbourhood. */
 #define VonNeumannR2                                                                                                   \
     Moore, {0, -2}, {0, 2}, {2, 0}, { -2, 0 }
+/** Triple Moore neighbourhood. */
 #define TripleMoore                                                                                                    \
     VonNeumannR2, {-1, -2}, {1, -2}, {-1, 2}, {1, 2}, {-2, -1}, {-2, 1}, {2, -1}, { 2, 1 }
+/** Triple Moore corners neighbourhood. */
 #define TripleMooreCorner                                                                                              \
     TripleMoore, {-2, -2}, {-2, 2}, {2, -2}, { 2, 2 }
 
-/* EXTERNAL NEIGHBOURHOOD CONSTANTS */
 extern const Neighbourhood VON_NEUMANN;
 extern const Neighbourhood VON_NEUMANN_CORNERS;
 extern const Neighbourhood LESSE;
@@ -44,11 +51,11 @@ extern const Neighbourhood VON_NEUMANN_R2;
 extern const Neighbourhood TRIPLE_MOORE;
 extern const Neighbourhood TRIPLE_MOORE_CORNER;
 
-Coordinate translate(Coordinate coord, int x, int y);
-void translate_coordinates(Coordinate *coords, unsigned int len, int x, int y);
+Coordinate translate(Coordinate coord, int32_t x, int32_t y);
+void translate_coordinates(Coordinate *coords, size_t len, int32_t x, int32_t y);
 Coordinate wrap(Environment const *env, Coordinate coord);
-bool *neighbours(Environment const *env, unsigned int x, unsigned int y, Neighbourhood const *neighbourhood,
+bool *neighbours(Environment const *env, uint32_t x, uint32_t y, Neighbourhood const *neighbourhood,
                  bool *neighbour_states);
-unsigned int num_neighbours(Environment const *env, unsigned int x, unsigned int y, Neighbourhood const *neighbourhood);
+unsigned int num_neighbours(Environment const *env, uint32_t x, uint32_t y, Neighbourhood const *neighbourhood);
 
 #endif // CONWAY_NEIGHBOURHOODS_H
